add fixes

+++ smarth-health.WebApi/Controllers/TimeLineItemController.cs
@@ -1,4 +1,5 @@
 ﻿// TimeLineItemsController.cs
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using smarth_health.WebApi.Models;
 using smarth_health.WebApi.Repositories;
@@ -17,13 +18,16 @@ namespace smarth_health.WebApi.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<IActionResult> GetAll()
         {
             var items = await _timeLineItemRepository.GetAllAsync();
             return Ok(items);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:guid}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> GetById(Guid id)
         {
             var item = await _timeLineItemRepository.GetByIdAsync(id);
@@ -34,6 +38,8 @@ namespace smarth_health.WebApi.Controllers
         }
 
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public async Task<IActionResult> Create(TimeLineItem item)
         {
             item.ID = Guid.NewGuid();
@@ -42,20 +48,33 @@ namespace smarth_health.WebApi.Controllers
             return CreatedAtAction(nameof(GetById), new { id = item.ID }, item);
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:guid}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> Update(Guid id, TimeLineItem item)
         {
             if (id != item.ID)
                 return BadRequest();
+
+            var existingItem = await _timeLineItemRepository.GetByIdAsync(id);
+            if (existingItem == null)
+                return NotFound();
 
             await _timeLineItemRepository.UpdateAsync(item);
 
             return NoContent();
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:guid}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var existingItem = await _timeLineItemRepository.GetByIdAsync(id);
+            if (existingItem == null)
+                return NotFound();
+
             await _timeLineItemRepository.DeleteAsync(id);
 
             return NoContent();
